Install specific Bundler version
Previously, the latest Bundler was installed. At the time of the commit,
that version was incompatible with the 3.x series of Redmine and had to
be replaced with an older version.

The version must be manually matched to the given Redmine version by
looking at the Gemfile of the desired Redmine. For example, Redmine
3.1.7 requires Bundler 1.5.0 or higher.